requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
idna==2.8
joblib==0.13.2
kafka-python==1.4.6
numpy==1.17.0
oauthlib==3.1.0
pycodestyle==2.5.0
PySocks==1.7.0
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
tweepy==3.8.0
urllib3==1.25.3
